requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy>=2.0.36,<3.0
Werkzeug==3.0.1
python-dotenv==1.0.0
reportlab==4.0.4
requests==2.32.5
Flask-Mail==0.10.0
Authlib==1.6.6
itsdangerous==2.2.0
psycopg2-binary==2.9.11
gunicorn==21.2.0
cryptography==41.0.7
PyMySQL==1.1.0
